Refactor + Rearrange
Selected MATLAB examples now moved into the doc folder
MAIN_pos_raytrace is very close to becoming the basis for the digital twin, so the bare minimum was migrated to the main level.

DOA and Positioning is possible via the 802.11az standard. Still a lot of work left in modeling the channel itself, but so far so good

There might be some problems with base_doa_pos, since it's only running in its current state. If we change something, there's a chance that it'll error out -- but it's more than likely that it's an inclusion issue. I just pasted the missing file into the +ranging folder, then wrote 'ranging.FILENAME' w/o the .m suffix and it'll work.